+++ apps/web/src/store/actions/fetchForGroup.js
@@ -51,6 +51,7 @@ const queryFragment = `group(slug: $slug, updateLastViewed: $updateLastViewed) {
     icon
     homeRoute
     menuViewCount
+    moreSpacesCount
     acceptedPostTypes
     typeDescriptor
     typeDescriptorPlural