+++ apps/web/src/store/actions/fetchGroupSpaces.js
@@ -10,6 +10,7 @@ export default function fetchGroupSpaces (groupId) {
       query: `query FetchGroupSpaces ($groupId: ID) {
         group(id: $groupId) {
           id
+          moreSpacesCount
           spaces {
             items {
               id
@@ -47,6 +48,7 @@ export default function fetchGroupSpaces (groupId) {
               active
               homeRoute
               menuViewCount
+              moreSpacesCount
               groupViews {
                 items {
                   id