+++ apps/web/src/store/actions/fetchGroupsMenuData.js
@@ -64,6 +64,7 @@ query FetchGroupsMenuData (
       icon
       homeRoute
       menuViewCount
+      moreSpacesCount
       visibility
       accessibility
       acceptedPostTypes
@@ -104,6 +105,7 @@ query FetchGroupsMenuData (
             icon
             homeRoute
             menuViewCount
+            moreSpacesCount
             description
             purpose
             location
@@ -211,6 +213,7 @@ query FetchGroupsMenuData (
           active
           homeRoute
           menuViewCount
+          moreSpacesCount
           groupViews {
             items {
               id